♻️ Refactor test structure into module folders
