+++ test/test_bin/test_delete_cnames.py
