+++ test/test_bin/test_low_github_licences.py
