+++ test/test_bin/test_report_on_inactive_users.py
